feat: split postmaster into postmaster-base and postmaster
Allows us to define different functionality for handling bit timing
changes in bootloader and apps.
