+++ libs/postmaster/meson.build
@@ -1,0 +1,5 @@
+postmaster_src = files('src' / 'postmaster.c')
+
+postmaster_dep = declare_dependency(sources: postmaster_src)
+
+postmaster_tidy_files = postmaster_src
